+++ client-and-server/server/node/server.js
@@ -1,44 +1,44 @@
-const express = require("express");
+const express = require('express');
 const app = express();
-const { resolve } = require("path");
+const { resolve } = require('path');
 // Copy the .env.example in the root into a .env file in this folder
-require("dotenv").config({ path: "./.env" });
-const stripe = require("stripe")(process.env.STRIPE_SECRET_KEY);
+require('dotenv').config({ path: './.env' });
+const stripe = require('stripe')(process.env.STRIPE_SECRET_KEY);
 
 app.use(express.static(process.env.STATIC_DIR));
 app.use(
   express.json({
     // We need the raw body to verify webhook signatures.
     // Let's compute it only when hitting the Stripe webhook endpoint.
-    verify: function(req, res, buf) {
-      if (req.originalUrl.startsWith("/webhook")) {
+    verify: function (req, res, buf) {
+      if (req.originalUrl.startsWith('/webhook')) {
         req.rawBody = buf.toString();
       }
-    }
+    },
   })
 );
 
-app.get("/", (req, res) => {
-  const path = resolve(process.env.STATIC_DIR + "/index.html");
+app.get('/', (req, res) => {
+  const path = resolve(process.env.STATIC_DIR + '/index.html');
   res.sendFile(path);
 });
 
-app.get("/config", (req, res) => {
+app.get('/config', (req, res) => {
   res.send({
     publicKey: process.env.STRIPE_PUBLISHABLE_KEY,
     basePrice: process.env.BASE_PRICE,
-    currency: process.env.CURRENCY
+    currency: process.env.CURRENCY,
   });
 });
 
 // Fetch the Checkout Session to display the JSON result on the success page
-app.get("/checkout-session", async (req, res) => {
+app.get('/checkout-session', async (req, res) => {
   const { sessionId } = req.query;
   const session = await stripe.checkout.sessions.retrieve(sessionId);
   res.send(session);
 });
 
-app.post("/create-checkout-session", async (req, res) => {
+app.post('/create-checkout-session', async (req, res) => {
   const domainURL = process.env.DOMAIN;
 
   const { quantity, locale } = req.body;
@@ -50,36 +50,36 @@ app.post("/create-checkout-session", async (req, res) => {
   // [customer_email] - lets you prefill the email input in the form
   // For full details see https://stripe.com/docs/api/checkout/sessions/create
   const session = await stripe.checkout.sessions.create({
-    payment_method_types: process.env.PAYMENT_METHODS.split(", "),
+    payment_method_types: process.env.PAYMENT_METHODS.split(', '),
     locale: locale,
     line_items: [
       {
-        name: "Pasha photo",
-        images: ["https://picsum.photos/300/300?random=4"],
+        name: 'Pasha photo',
+        images: ['https://picsum.photos/300/300?random=4'],
         quantity: quantity,
         currency: process.env.CURRENCY,
-        amount: process.env.BASE_PRICE // Keep the amount on the server to prevent customers from manipulating on client
-      }
+        amount: process.env.BASE_PRICE, // Keep the amount on the server to prevent customers from manipulating on client
+      },
     ],
     // ?session_id={CHECKOUT_SESSION_ID} means the redirect will have the session ID set as a query param
     success_url: `${domainURL}/success.html?session_id={CHECKOUT_SESSION_ID}`,
-    cancel_url: `${domainURL}/canceled.html`
+    cancel_url: `${domainURL}/canceled.html`,
   });
 
   res.send({
-    sessionId: session.id
+    sessionId: session.id,
   });
 });
 
 // Webhook handler for asynchronous events.
-app.post("/webhook", async (req, res) => {
+app.post('/webhook', async (req, res) => {
   let data;
   let eventType;
   // Check if webhook signing is configured.
   if (process.env.STRIPE_WEBHOOK_SECRET) {
     // Retrieve the event by verifying the signature using the raw body and secret.
     let event;
-    let signature = req.headers["stripe-signature"];
+    let signature = req.headers['stripe-signature'];
 
     try {
       event = stripe.webhooks.constructEvent(
@@ -101,7 +101,7 @@ app.post("/webhook", async (req, res) => {
     eventType = req.body.type;
   }
 
-  if (eventType === "checkout.session.completed") {
+  if (eventType === 'checkout.session.completed') {
     console.log(`🔔  Payment received!`);
   }
